add seeder bagian gre

+++ database/migrations/2024_03_21_140534_create_detail_hampers_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('hampers_id')->constrained('hampers');
             $table->foreignId('produk_id')->constrained('produks');
-            $table->integer('jumlah_produk');
+            $table->float('jumlah_produk');
             $table->timestamps();
         });
     }
